refactor: put user, board, column and task into resources directory
