+++ src/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
-import { UserModule } from 'src/user/user.module';
+import { UserModule } from 'src/resources/user/user.module';
 import { AuthController } from './auth.controller';
 import 'dotenv/config';
 import { JwtStrategy } from './jwt.strategy';
