+++ src/resources/board/board.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ColumnModule } from 'src/column/column.module';
-import { TaskModule } from 'src/task/task.module';
+import { ColumnModule } from 'src/resources/column/column.module';
+import { TaskModule } from 'src/resources/task/task.module';
 import { BoardController } from './board.controller';
 import { BoardService } from './board.service';
 import { Board } from './models/board.entity';
