+++ src/resources/board/models/board-update.dto.ts
@@ -1,6 +1,6 @@
 import { Type } from 'class-transformer';
 import { IsArray, IsDefined, IsString, ValidateNested } from 'class-validator';
-import { Column } from 'src/column/column.entity';
+import { Column } from 'src/resources/column/column.entity';
 
 export class BoardUpdateDto {
   @IsString()
