+++ src/resources/user/models/user.entity.ts
